+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `buildDateTimeColumn` helper
 - Add entity access check to query
+- Add abstract base class for actions with batch processing
 
 ## [1.7.1] - 2021-03-22
 ### Changed